+++ Changes.md
@@ -5,6 +5,7 @@ Fixes
 -----
 
 - UIEditor : Fixed Python 3 compatibility in plug presets editor.
+- NodeGadget : Fixed intermittent shutdown crash.
 
 API
 ---
@@ -233,6 +234,14 @@ Build
   - USD 20.11
   - OpenSSL 1.1.1h
   - See https://github.com/GafferHQ/dependencies/releases/tag/2.1.1 for full details.
+
+0.58.6.x (relative to 0.58.6.1)
+========
+
+Fixes
+-----
+
+- NodeGadget : Fixed intermittent shutdown crash.
 
 0.58.6.1 (relative to 0.58.6.0)
 ========